requirements.txt
requests>=2.31.0
beautifulsoup4>=4.8.1
streamlit>=1.44.0
googlemaps>=4.2.0
google-ai-generativelanguage
pandas>=2.2.3
plotly>=5.18.0